+++ cedar-policy-symcc/src/lib.rs
@@ -157,6 +157,11 @@ impl CompiledPolicy {
         self.policy.effect()
     }
 
+    /// Get the (post-typecheck) `Policy` that this `CompiledPolicy` represents
+    pub fn policy(&self) -> &cedar_policy_core::ast::Policy {
+        &self.policy.policy
+    }
+
     /// Convert a `CompiledPolicy` to a `CompiledPolicies` representing a
     /// singleton policyset with just that policy.
     ///
@@ -187,6 +192,11 @@ impl CompiledPolicies {
         Ok(Self {
             policies: symccopt::CompiledPolicies::compile(pset.as_ref(), env, schema)?,
         })
+    }
+
+    /// Get the (post-typecheck) `PolicySet` that this `CompiledPolicies` represents
+    pub fn policies(&self) -> &cedar_policy_core::ast::PolicySet {
+        &self.policies.policies
     }
 }